Move resolveTables into BucketStorageBatch, where it belongs.
